--- src/main/java/lawbringer/LawbringerException.java
@@ -1,10 +1,0 @@
-package lawbringer;
-
-/**
- * This class represents a custom exception to use in the Lawbringer class.
- */
-public class LawbringerException extends Exception {
-    public LawbringerException(String exception) {
-        super(exception);
-    }
-}
