+++ src/main/java/uhtredragnarson/Deadline.java
@@ -1,4 +1,4 @@
-package lawbringer;
+package uhtredragnarson;
 
 /**
  * The Deadline class extends the Task class and represents a deadline that has a specific
