+++ src/main/java/uhtredragnarson/Event.java
@@ -1,4 +1,4 @@
-package lawbringer;
+package uhtredragnarson;
 
 /**
  * The Event class extends the Task class and represents an event that begin and
