+++ src/main/java/uhtredragnarson/Task.java
@@ -1,4 +1,4 @@
-package lawbringer;
+package uhtredragnarson;
 
 /**
  * This is an abstract class in which the ToDo, Deadline and Event class will extend from.
