+++ src/main/java/uhtredragnarson/UhtredRagnarsonException.java
@@ -1,0 +1,10 @@
+package uhtredragnarson;
+
+/**
+ * This class represents a custom exception to use in the UhtredRagnarson class.
+ */
+public class UhtredRagnarsonException extends Exception {
+    public UhtredRagnarsonException(String exception) {
+        super(exception);
+    }
+}
